Update Makefile (#26)
This PR adds two new targets to the Makefile: clean-manifest and rebranding.

clean-manifest removes all comments from the manifest.yaml file, making it easier to read.

rebranding prompts the user to enter a new package ID and title, and updates all relevant files in the repository with the new values. It also comments out the rebranding code in the Makefile to prevent accidental usage in the future.

These targets can be run with the command make clean-manifest and make rebranding, respectively.